cv-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic-settings>=2.2.0
opencv-python-headless>=4.9.0
Pillow>=10.3.0
numpy>=1.26.0
python-multipart>=0.0.9
requests>=2.31.0
torch>=2.2.0
torchvision>=0.17.0
